Move recorded tracks by identity on close (crashing)

+++ zoo_vision_ros2/src/zoo_vision/src/identifier.cpp
@@ -260,7 +260,7 @@ void Identifier::onDetection(zoo_msgs::msg::Detection &msg, const torch::Tensor 
       continue;
     }
 
-    if (identity >= identityCount + 1) {
+    if (uint32_t(identity) >= identityCount + 1) {
       throw std::runtime_error(std::format("identity out of range, {} >= {}", identity, identityCount));
     }
